Group roles and move sources according to the new grouping.
The roles are to be grouped according to the feature group to which
they belong. The initial ones are:

* base for things which are useful on even most minimal setups;
* dev for development tools;
* gui for software which runs with graphical interface.

If some software piece logically belongs to multiple groups (e.g.
Netbeans belongs to both dev/java and gui groups), the most specific
group should be used for determining where the role has to go to.