Add managed drop-ins support for .profile and .bashrc.
There is currenly a support for unmanaged drop-ins, but it would be
beneficial to also support managed drop-ins to allow roles inserting
their drop-ins too.

Directories for managed and unmanaged drop-ins are segregated to allow
safely cleaning managed drop-ins in order to remove deprecated drop-ins
when newer version of playbooks is applied.