Add fzf vim plugin.
Fzf ships its own set of functions in its main repository, and fzf.vim
relies on that functions. On Arch, these definitions are shipped with
the package, but when we install fzf as a standalone binary, we have to
install these definitions manually. The easiest way is to treat fzf as a
separate plugin, as the required definitions are stored in `plugin`
directory inside the repo and, thus, are recognizable by Pathogen as a
plugin.