Replace YCM custom installation process with default install.py script.
It is a heavy burden to support the plugin installation, as installation
guide is now unmaintained and plugin developers recommend to use the
installation script as the only true way to get YCM up and running.

The only reason to avoid this script currently is the need to customize
the Rust support installation. To do this, we don't need to install
everything manually; instead, we may install the plugin as recommended
but without Rust support, and then configure Rust support manually.