Create fd and bat symlinks for PureOS.
Debian-like distros use custom names for these binaries to avoid name
clash with other packages, and we need to handle this somehow, as we
already depend on these names in configs and scripts.

Given that we don't need tools which caused this workaround in PATH, we
may just create symlinks in some PATH directory (/usr/local/bin seems
the best option here) to renamed binaries and pretend that we have no
issue.